Make test suit logic free-standing functions
